lir,wasm,docs: the prose gate reads a stamp as a claim, so meet it or drop it
`tests/integration/prose.rs` checks what a machine can decide about every
stamped file, and the stamps this branch added opted eleven files into checks
they did not meet.

Three matches over the instruction set split where the budget says they must:
`emit/instr/ops.rs` hands its tail to `ops/intrinsics.rs`,
`wasm/instruction/dispatch.rs` to `instruction/intrinsics.rs`, and the
arithmetic-opcode tests leave `emit/tests.rs` for `emit/tests/opcodes.rs`.
Five call-outs came under 140 characters, three file names in stamped
documents became links, and two uses of a banned word were rewritten.

`types/instr.rs` is one enum, which Rust gives no way to split, so it carries
no stamp and stays in the queue — the policy's own answer for a file that
breaks a rule. Bringing it inside the budget means nesting a variant group
into a sub-enum and rewriting every exhaustive match in the crate, which is
its own change. The two module indexes record where each file stands.

The earlier claim that a dispatch table may run to 800 lines does not survive
the build: the check is a flat 500 over every stamped file.
